razpai/requirements.txt
RPi.GPIO
dht11
smbus2
cryptography==2.8
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth==1.10.0
google-cloud-iot==0.3.0
google-cloud-pubsub==1.1.0
pyjwt==1.7.1
requests==2.22.0